+++ packages/gatsby/src/commands/build.ts
@@ -30,7 +30,6 @@ import {
   showFeedbackRequest,
   showSevenDayFeedbackRequest,
 } from "../utils/feedback"
-import * as buildUtils from "./build-utils"
 import { actions } from "../redux/actions"
 import { waitUntilAllJobsComplete } from "../utils/wait-until-jobs-complete"
 import { Stage } from "./types"
@@ -240,7 +239,7 @@ module.exports = async function build(program: IBuildArgs): Promise<void> {
   workerPool.end()
   buildActivity.end()
 
-  if (process.argv.includes(`--log-pages`)) {
+  if (program.logPages) {
     if (toRegenerate.length) {
       report.info(
         `Built pages:\n${toRegenerate
@@ -258,7 +257,7 @@ module.exports = async function build(program: IBuildArgs): Promise<void> {
     }
   }
 
-  if (process.argv.includes(`--write-to-file`)) {
+  if (program.writeToFile) {
     const createdFilesPath = path.resolve(
       `${program.directory}/.cache`,
       `newPages.txt`