+++ packages/gatsby/src/utils/webpack-utils.ts
@@ -59,6 +59,12 @@ type CSSModulesOptions =
       exportOnlyLocals?: boolean
     }
 
+type MiniCSSExtractLoaderModuleOptions =
+  | undefined
+  | boolean
+  | {
+      namedExport?: boolean
+    }
 /**
  * Utils that produce webpack `loader` objects
  */
@@ -225,10 +231,29 @@ export const createWebpackUtils = (
       }
     },
 
-    miniCssExtract: (options = {}) => {
+    miniCssExtract: (
+      options: {
+        modules?: MiniCSSExtractLoaderModuleOptions
+      } = {}
+    ) => {
+      let moduleOptions: MiniCSSExtractLoaderModuleOptions = undefined
+
+      const { modules, ...restOptions } = options
+
+      if (typeof modules === `boolean` && options.modules) {
+        moduleOptions = {
+          namedExport: true,
+        }
+      } else {
+        moduleOptions = modules
+      }
+
       return {
         loader: MiniCssExtractPlugin.loader,
-        options,
+        options: {
+          modules: moduleOptions,
+          ...restOptions,
+        },
       }
     },
 
@@ -236,7 +261,8 @@ export const createWebpackUtils = (
       let modulesOptions: CSSModulesOptions = false
       if (options.modules) {
         modulesOptions = {
-          auto: false,
+          auto: undefined,
+          namedExport: true,
           localIdentName: `[name]--[local]--[hash:base64:5]`,
           exportLocalsConvention: `dashesOnly`,
           exportOnlyLocals: isSSR,
@@ -391,9 +417,6 @@ export const createWebpackUtils = (
     } = {}): RuleSetRule => {
       return {
         test: /\.(js|mjs|jsx)$/,
-        resolve: {
-          fullySpecified: false,
-        },
         include: (modulePath: string): boolean => {
           // when it's not coming from node_modules we treat it as a source file.
           if (!vendorRegex.test(modulePath)) {
@@ -487,9 +510,6 @@ export const createWebpackUtils = (
 
       return {
         test: /\.(js|mjs)$/,
-        resolve: {
-          fullySpecified: false,
-        },
         exclude: (modulePath: string): boolean => {
           // If dep is user land code, exclude
           if (!vendorRegex.test(modulePath)) {
@@ -593,7 +613,7 @@ export const createWebpackUtils = (
     const css: IRuleUtils["css"] = (options = {}): RuleSetRule => {
       const { browsers, ...restOptions } = options
       const use = [
-        loaders.miniCssExtract(),
+        loaders.miniCssExtract(restOptions),
         loaders.css({ ...restOptions, importLoaders: 1 }),
         loaders.postcss({ browsers }),
       ]