+++ packages/gatsby/src/utils/webpack.config.js
@@ -18,8 +18,8 @@ import { createWebpackUtils } from "./webpack-utils"
 import { hasLocalEslint } from "./local-eslint-config-finder"
 import { getAbsolutePathForVirtualModule } from "./gatsby-webpack-virtual-modules"
 import { StaticQueryMapper } from "./webpack/static-query-mapper"
-import { TmpMiniCssExtractContentHashOverWrite } from "./webpack/tmp-mini-css-extract-contenthash-overwrite"
 import { getBrowsersList } from "./browserslist"
+import { builtinModules } from "module"
 
 const FRAMEWORK_BUNDLES = [`react`, `react-dom`, `scheduler`, `prop-types`]
 
@@ -172,13 +172,7 @@ module.exports = async (
       case `develop`:
         return {
           polyfill: directoryPath(`.cache/polyfill-entry`),
-          commons: [
-            process.env.GATSBY_HOT_LOADER !== `fast-refresh` &&
-              `${require.resolve(
-                `webpack-hot-middleware/client`
-              )}?path=${getHmrPath()}`,
-            directoryPath(`.cache/app`),
-          ].filter(Boolean),
+          commons: [directoryPath(`.cache/app`)],
         }
       case `develop-html`:
         return {
@@ -222,8 +216,7 @@ module.exports = async (
       case `develop`:
         configPlugins = configPlugins
           .concat([
-            process.env.GATSBY_HOT_LOADER === `fast-refresh` &&
-              plugins.fastRefresh(),
+            plugins.fastRefresh(),
             plugins.hotModuleReplacement(),
             plugins.noEmitOnErrors(),
             plugins.eslintGraphqlSchemaReload(),
@@ -248,8 +241,6 @@ module.exports = async (
             filename: `[name].[contenthash].css`,
             chunkFilename: `[name].[contenthash].css`,
           }),
-          // remove after https://github.com/webpack-contrib/mini-css-extract-plugin/issues/701
-          new TmpMiniCssExtractContentHashOverWrite(),
           // Write out stats object mapping named dynamic imports (aka page
           // components) to all their async chunks.
           plugins.extractStats(),
@@ -265,9 +256,7 @@ module.exports = async (
   function getDevtool() {
     switch (stage) {
       case `develop`:
-        return process.env.GATSBY_HOT_LOADER !== `fast-refresh`
-          ? `cheap-module-source-map`
-          : `eval-cheap-module-source-map`
+        return `eval-cheap-module-source-map`
       // use a normal `source-map` for the html phases since
       // it gives better line and column numbers
       case `develop-html`:
@@ -296,18 +285,35 @@ module.exports = async (
     // Common config for every env.
     // prettier-ignore
     let configRules = [
-      rules.js({
-        modulesThatUseGatsby,
-      }),
-      // Webpack expects extensions when importing to mimic ESM spec.
+      // Webpack expects extensions when importing ESM modules as that's what the spec describes.
       // Not all libraries have adapted so we don't enforce its behaviour
       // @see https://github.com/webpack/webpack/issues/11467
       {
-        test: /\.m?js/,
+        test: /\.mjs$/i,
         resolve: {
-            fullySpecified: false
+          byDependency: {
+            esm: {
+              fullySpecified: false
+            }
+          }
         }
       },
+      {
+        test: /\.js$/i,
+        descriptionData: {
+          type: `module`
+        },
+        resolve: {
+          byDependency: {
+            esm: {
+              fullySpecified: false
+            }
+          }
+        }
+      },
+      rules.js({
+        modulesThatUseGatsby,
+      }),
       rules.yaml(),
       rules.fonts(),
       rules.images(),
@@ -350,10 +356,7 @@ module.exports = async (
         }
 
         // Enforce fast-refresh rules even with local eslint config
-        if (
-          isCustomEslint &&
-          process.env.GATSBY_HOT_LOADER === `fast-refresh`
-        ) {
+        if (isCustomEslint) {
           configRules = configRules.concat([rules.eslintRequired()])
         }
 
@@ -362,19 +365,6 @@ module.exports = async (
             oneOf: [rules.cssModules(), rules.css()],
           },
         ])
-
-        // RHL will patch React, replace React-DOM by React-🔥-DOM and work with fiber directly
-        // It's necessary to remove the warning in console (https://github.com/gatsbyjs/gatsby/issues/11934)
-        // TODO: Remove entire block when we make fast-refresh the default
-        if (process.env.GATSBY_HOT_LOADER !== `fast-refresh`) {
-          configRules.push({
-            include: /node_modules\/react-dom/,
-            test: /\.jsx?$/,
-            use: {
-              loader: require.resolve(`./webpack-hmr-hooks-patch`),
-            },
-          })
-        }
 
         break
       }
@@ -433,12 +423,6 @@ module.exports = async (
         // relative path imports are used sometimes
         // See https://stackoverflow.com/a/49455609/6420957 for more details
         "@babel/runtime": getPackageRoot(`@babel/runtime`),
-        // TODO: Remove entire block when we make fast-refresh the default
-        ...(process.env.GATSBY_HOT_LOADER !== `fast-refresh`
-          ? {
-              "react-hot-loader": getPackageRoot(`react-hot-loader`),
-            }
-          : {}),
         "react-lifecycles-compat": directoryPath(
           `.cache/react-lifecycles-compat.js`
         ),
@@ -506,7 +490,6 @@ module.exports = async (
       hints: false,
     },
     mode: getMode(),
-    cache: false,
 
     resolveLoader: getResolveLoader(),
     resolve: getResolve(stage),
@@ -521,9 +504,46 @@ module.exports = async (
     )}`
   }
 
+  const isCssModule = module => module.type === `css/mini-extract`
+  if (stage === `develop`) {
+    config.optimization = {
+      splitChunks: {
+        chunks: `all`,
+        cacheGroups: {
+          default: false,
+          defaultVendors: false,
+          framework: {
+            chunks: `all`,
+            name: `framework`,
+            // This regex ignores nested copies of framework libraries so they're bundled with their issuer.
+            test: new RegExp(
+              `(?<!node_modules.*)[\\\\/]node_modules[\\\\/](${FRAMEWORK_BUNDLES.join(
+                `|`
+              )})[\\\\/]`
+            ),
+            priority: 40,
+            // Don't let webpack eliminate this chunk (prevents this chunk from becoming a part of the commons chunk)
+            enforce: true,
+          },
+          // Bundle all css & lazy css into one stylesheet to make sure lazy components do not break
+          // TODO make an exception for css-modules
+          styles: {
+            test(module) {
+              return isCssModule(module)
+            },
+
+            name: `commons`,
+            priority: 40,
+            enforce: true,
+          },
+        },
+      },
+      minimizer: [],
+    }
+  }
+
   if (stage === `build-javascript`) {
     const componentsCount = store.getState().components.size
-    const isCssModule = module => module.type === `css/mini-extract`
 
     const splitChunks = {
       chunks: `all`,
@@ -640,7 +660,7 @@ module.exports = async (
     // removes node internals from bundle
     // https://webpack.js.org/configuration/externals/#externalspresets
     config.externalsPresets = {
-      node: true,
+      node: stage === `build-html` ? false : true,
     }
 
     // Packages we want to externalize to save some build time
@@ -672,20 +692,8 @@ module.exports = async (
         // User modules that do not need to be part of the bundle
         if (userExternalList.some(item => checkItem(item, request))) {
           // TODO figure out to make preact work with this too
-          let modifiedRequest = request
-          if (
-            stage === `develop-html` &&
-            isCI() &&
-            process.env.GATSBY_EXPERIMENTAL_DEV_SSR
-          ) {
-            if (request === `react`) {
-              modifiedRequest = `react/cjs/react.production.min.js`
-            } else if (request === `react-dom/server`) {
-              modifiedRequest = `react-dom/cjs/react-dom-server.node.production.min.js`
-            }
-          }
 
-          resolver(context, modifiedRequest, (err, newRequest) => {
+          resolver(context, request, (err, newRequest) => {
             if (err) {
               callback(err)
               return
@@ -712,6 +720,34 @@ module.exports = async (
         callback()
       },
     ]
+
+    if (stage === `build-html`) {
+      const builtinModulesToTrack = [
+        `fs`,
+        `http`,
+        `http2`,
+        `https`,
+        `child_process`,
+      ]
+      const builtinsExternalsDictionary = builtinModules.reduce(
+        (acc, builtinModule) => {
+          if (builtinModulesToTrack.includes(builtinModule)) {
+            acc[builtinModule] = `commonjs ${path.join(
+              program.directory,
+              `.cache`,
+              `ssr-builtin-trackers`,
+              builtinModule
+            )}`
+          } else {
+            acc[builtinModule] = `commonjs ${builtinModule}`
+          }
+          return acc
+        },
+        {}
+      )
+
+      config.externals.unshift(builtinsExternalsDictionary)
+    }
   }
 
   if (stage === `develop`) {